Implement column based input component factory for number
