+++ mathesar_ui/src/component-library/number-input/index.ts
@@ -1,0 +1,3 @@
+export { default as NumberInput } from './NumberInput.svelte';
+export { default as StringifiedNumberInput } from './StringifiedNumberInput.svelte';
+export { StringifiedNumberFormatter } from './number-formatter';
