+++ mathesar_ui/src/components/cell/data-types/components/number/NumberCell.svelte
@@ -1,57 +1,32 @@
 <script lang="ts">
-  import { StringifiedNumberInput } from '@mathesar-component-library';
-  import type { NumberFormat } from '@mathesar/api/tables/columns';
-  import { StringifiedNumberFormatter } from '@mathesar/component-library/number-input/number-formatter';
+  import {
+    StringifiedNumberFormatter,
+    isDefinedNonNullable,
+  } from '@mathesar-component-library';
   import SteppedInputCell from '../SteppedInputCell.svelte';
   import type { NumberCellProps } from '../typeDefinitions';
+  import NumberCellInput from './NumberCellInput.svelte';
 
   type $$Props = NumberCellProps;
 
-  type ParentValue = $$Props['value'];
-  type ChildValue = string | null;
-
   export let isActive: $$Props['isActive'];
-  let parentValue: ParentValue = undefined;
-  export { parentValue as value };
+  export let value: $$Props['value'];
   export let disabled: $$Props['disabled'];
 
-  export let format: $$Props['format'];
-  // TODO connect this to StringifiedNumberInput
+  export let locale: $$Props['locale'];
   export let isPercentage: $$Props['isPercentage'];
 
-  // prettier-ignore
-  const localeMap = new Map<NumberFormat, string>([
-    ['english' , 'en'    ],
-    ['german'  , 'de'    ],
-    ['french'  , 'fr'    ],
-    ['hindi'   , 'hi'    ],
-    ['swiss'   , 'de-CH' ],
-  ]);
-
   $: formatterOptions = {
-    locale: (format && localeMap.get(format)) ?? undefined,
+    locale,
     allowFloat: true, // TODO set based on DB options
     allowNegative: true,
   };
   $: formatter = new StringifiedNumberFormatter(formatterOptions);
 
-  let childValue: ChildValue = null;
-
-  function getNewChildValue(newParentValue: ParentValue): ChildValue {
-    if (newParentValue === undefined || newParentValue === null) {
-      return null;
-    }
-    return String(newParentValue);
-  }
-  function handleParentValueChange(newParentValue: ParentValue) {
-    childValue = getNewChildValue(newParentValue);
-  }
-  $: handleParentValueChange(parentValue);
-
   function formatValue(
     v: string | number | null | undefined,
   ): string | null | undefined {
-    if (v === undefined || v === null) {
+    if (!isDefinedNonNullable(v)) {
       return v;
     }
     return formatter.format(String(v));
@@ -59,7 +34,7 @@
 </script>
 
 <SteppedInputCell
-  value={parentValue}
+  {value}
   {isActive}
   {disabled}
   {formatValue}
@@ -69,15 +44,12 @@
   on:activate
   on:update
 >
-  <StringifiedNumberInput
-    focusOnMount={true}
+  <NumberCellInput
     {disabled}
-    value={childValue}
+    bind:value
+    {...formatterOptions}
+    {isPercentage}
     on:blur={handleInputBlur}
     on:keydown={handleInputKeydown}
-    on:input={({ detail: newChildValue }) => {
-      parentValue = newChildValue;
-    }}
-    {...formatterOptions}
   />
 </SteppedInputCell>
